Use TCP connection to graphite.
Running this against our test graphite instance worked, but it failed
when run against our production graphite instance because the production
instance is behind a load balancer.

I suspect the CLI should probably have a flag to toggle between TCP and
UDP but TCP seems to work well enough for both test and prod use cases
in our infrastructure so I'll leave the CLI flag for someone who needs
it to implement.